// src/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type BaseCartLineEdge = {
  __typename?: 'BaseCartLineEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of BaseCartLineEdge. */
  node: BaseCartLine;
};
// ...
export type Cart = HasMetafields & Node & {
  __typename?: 'Cart';
  /** The gift cards that have been applied to the cart. */
  appliedGiftCards: Array<AppliedGiftCard>;
  /** An attribute associated with the cart. */
  attribute?: Maybe<Attribute>;
  /** The attributes associated with the cart. Attributes are represented as key-value pairs. */
  attributes: Array<Attribute>;
  /** Information about the buyer that's interacting with the cart. */
  buyerIdentity: CartBuyerIdentity;
  /** The URL of the checkout for the cart. */
  checkoutUrl: Scalars['URL']['output'];
  /** The estimated costs that the buyer will pay at checkout. The costs are subject to change and changes will be reflected at checkout. The `cost` field uses the `buyerIdentity` field to determine [international pricing](https://shopify.dev/custom-storefronts/internationalization/international-pricing). */
  cost: CartCost;
  /** The date and time when the cart was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The delivery properties of the cart. */
  delivery: CartDelivery;
  /**
   * The delivery groups available for the cart, based on the buyer identity default
   * delivery address preference or the default address of the logged-in customer.
   *
   */
  deliveryGroups: CartDeliveryGroupConnection;
  /** The discounts that have been applied to the entire cart. */
  discountAllocations: Array<CartDiscountAllocation>;
  /** The case-insensitive discount codes that the customer added at checkout. */
  discountCodes: Array<CartDiscountCode>;
  /**
   * The estimated costs that the buyer will pay at checkout. The estimated costs are subject to change and changes will be reflected at checkout. The `estimatedCost` field uses the `buyerIdentity` field to determine [international pricing](https://shopify.dev/custom-storefronts/internationalization/international-pricing).
   * @deprecated Use `cost` instead.
   */
  estimatedCost: CartEstimatedCost;
  /** A globally-unique ID. */
  id: Scalars['ID']['output'];
  /** A list of lines containing information about the items the customer intends to purchase. */
  lines: BaseCartLineConnection;
  /** A [custom field](https://shopify.dev/docs/apps/build/custom-data), including its `namespace` and `key`, that's associated with a Shopify resource for the purposes of adding and storing additional information. */
  metafield?: Maybe<Metafield>;
  /** A list of [custom fields](/docs/apps/build/custom-data) that a merchant associates with a Shopify resource. */
  metafields: Array<Maybe<Metafield>>;
  /** A note that's associated with the cart. For example, the note can be a personalized message to the buyer. */
  note?: Maybe<Scalars['String']['output']>;
  /** The total number of items in the cart. */
  totalQuantity: Scalars['Int']['output'];
  /** The date and time when the cart was updated. */
  updatedAt: Scalars['DateTime']['output'];
};
// ...
export type CartLineInput = {
  /**
   * An array of key-value pairs that contains additional information about the merchandise line.
   *
   * The input must not contain more than `250` values.
   */
  attributes?: InputMaybe<Array<AttributeInput>>;
  /** The ID of the merchandise that the buyer intends to purchase. */
  merchandiseId: Scalars['ID']['input'];
  /** The quantity of the merchandise. */
  quantity?: InputMaybe<Scalars['Int']['input']>;
  /** The ID of the selling plan that the merchandise is being purchased with. */
  sellingPlanId?: InputMaybe<Scalars['ID']['input']>;
};

/** The input fields to update a line item on a cart. */
export type CartLineUpdateInput = {
  /**
   * An array of key-value pairs that contains additional information about the merchandise line.
   *
   * The input must not contain more than `250` values.
   */
  attributes?: InputMaybe<Array<AttributeInput>>;
  /** The ID of the merchandise line. */
  id: Scalars['ID']['input'];
  /** The ID of the merchandise for the line item. */
  merchandiseId?: InputMaybe<Scalars['ID']['input']>;
  /** The quantity of the line item. */
  quantity?: InputMaybe<Scalars['Int']['input']>;
  /** The ID of the selling plan that the merchandise is being purchased with. */
  sellingPlanId?: InputMaybe<Scalars['ID']['input']>;
};
// ...
export type MoneyV2 = {
  __typename?: 'MoneyV2';
  /** Decimal money amount. */
  amount: Scalars['Decimal']['output'];
  /** Currency of the money. */
  currencyCode: CurrencyCode;
};
// ...
export enum ProductCollectionSortKeys {
  /** Sort by the `best-selling` value. */
  BestSelling = 'BEST_SELLING',
  /** Sort by the `collection-default` value. */
  CollectionDefault = 'COLLECTION_DEFAULT',
  /** Sort by the `created` value. */
  Created = 'CREATED',
  /** Sort by the `id` value. */
  Id = 'ID',
  /** Sort by the `manual` value. */
  Manual = 'MANUAL',
  /** Sort by the `price` value. */
  Price = 'PRICE',
  /**
   * Sort by relevance to the search terms when the `query` parameter is specified on the connection.
   * Don't use this sort key when no search query is specified.
   *
   */
  Relevance = 'RELEVANCE',
  /** Sort by the `title` value. */
  Title = 'TITLE'
}
// ...
export enum ProductSortKeys {
  /** Sort by the `best_selling` value. */
  BestSelling = 'BEST_SELLING',
  /** Sort by the `created_at` value. */
  CreatedAt = 'CREATED_AT',
  /** Sort by the `id` value. */
  Id = 'ID',
  /** Sort by the `price` value. */
  Price = 'PRICE',
  /** Sort by the `product_type` value. */
  ProductType = 'PRODUCT_TYPE',
  /**
   * Sort by relevance to the search terms when the `query` parameter is specified on the connection.
   * Don't use this sort key when no search query is specified.
   *
   */
  Relevance = 'RELEVANCE',
  /** Sort by the `title` value. */
  Title = 'TITLE',
  /** Sort by the `updated_at` value. */
  UpdatedAt = 'UPDATED_AT',
  /** Sort by the `vendor` value. */
  Vendor = 'VENDOR'
}
// ...
export type CartFragment = { __typename?: 'Cart', id: string, checkoutUrl: any, totalQuantity: number, cost: { __typename?: 'CartCost', subtotalAmount: { __typename?: 'MoneyV2', amount: any, currencyCode: CurrencyCode }, totalAmount: { __typename?: 'MoneyV2', amount: any, currencyCode: CurrencyCode }, totalTaxAmount?: { __typename?: 'MoneyV2', amount: any, currencyCode: CurrencyCode } | null }, lines: { __typename?: 'BaseCartLineConnection', edges: Array<{ __typename?: 'BaseCartLineEdge', node: { __typename?: 'CartLine', id: string, quantity: number, cost: { __typename?: 'CartLineCost', totalAmount: { __typename?: 'MoneyV2', amount: any, currencyCode: CurrencyCode } }, merchandise: { __typename?: 'ProductVariant', quantityAvailable?: number | null, id: string, title: string, selectedOptions: Array<{ __typename?: 'SelectedOption', name: string, value: string }>, product: (
            { __typename?: 'Product' }
            & { ' $fragmentRefs'?: { 'ProductFragment': ProductFragment } }
          ) } } | { __typename?: 'ComponentizableCartLine', id: string, quantity: number, cost: { __typename?: 'CartLineCost', totalAmount: { __typename?: 'MoneyV2', amount: any, currencyCode: CurrencyCode } }, merchandise: { __typename?: 'ProductVariant', quantityAvailable?: number | null, id: string, title: string, selectedOptions: Array<{ __typename?: 'SelectedOption', name: string, value: string }>, product: (
            { __typename?: 'Product' }
            & { ' $fragmentRefs'?: { 'ProductFragment': ProductFragment } }
          ) } } }> } } & { ' $fragmentName'?: 'CartFragment' };

export type ImageFragment = { __typename?: 'Image', url: any, altText?: string | null, width?: number | null, height?: number | null } & { ' $fragmentName'?: 'ImageFragment' };

export type ProductFragment = { __typename?: 'Product', id: string, handle: string, availableForSale: boolean, title: string, description: string, descriptionHtml: any, tags: Array<string>, updatedAt: any, shortDescription?: { __typename?: 'Metafield', value: string } | null, options: Array<{ __typename?: 'ProductOption', id: string, name: string, values: Array<string> }>, priceRange: { __typename?: 'ProductPriceRange', maxVariantPrice: { __typename?: 'MoneyV2', amount: any, currencyCode: CurrencyCode }, minVariantPrice: { __typename?: 'MoneyV2', amount: any, currencyCode: CurrencyCode } }, variants: { __typename?: 'ProductVariantConnection', edges: Array<{ __typename?: 'ProductVariantEdge', node: { __typename?: 'ProductVariant', quantityAvailable?: number | null, id: string, title: string, availableForSale: boolean, selectedOptions: Array<{ __typename?: 'SelectedOption', name: string, value: string }>, price: { __typename?: 'MoneyV2', amount: any, currencyCode: CurrencyCode }, compareAtPrice?: { __typename?: 'MoneyV2', amount: any, currencyCode: CurrencyCode } | null, image?: { __typename?: 'Image', url: any, altText?: string | null } | null } }> }, featuredImage?: (
    { __typename?: 'Image' }
    & { ' $fragmentRefs'?: { 'ImageFragment': ImageFragment } }
  ) | null, images: { __typename?: 'ImageConnection', edges: Array<{ __typename?: 'ImageEdge', node: (
        { __typename?: 'Image' }
        & { ' $fragmentRefs'?: { 'ImageFragment': ImageFragment } }
      ) }> }, seo: (
    { __typename?: 'SEO' }
    & { ' $fragmentRefs'?: { 'SeoFragment': SeoFragment } }
  ) } & { ' $fragmentName'?: 'ProductFragment' };

export type SeoFragment = { __typename?: 'SEO', description?: string | null, title?: string | null } & { ' $fragmentName'?: 'SeoFragment' };

export type AddToCartMutationVariables = Exact<{
  cartId: Scalars['ID']['input'];
  lines: Array<CartLineInput> | CartLineInput;
}>;


export type AddToCartMutation = { __typename?: 'Mutation', cartLinesAdd?: { __typename?: 'CartLinesAddPayload', cart?: (
      { __typename?: 'Cart' }
      & { ' $fragmentRefs'?: { 'CartFragment': CartFragment } }
    ) | null } | null };

export type CreateCartMutationVariables = Exact<{
  lineItems?: InputMaybe<Array<CartLineInput> | CartLineInput>;
}>;


export type CreateCartMutation = { __typename?: 'Mutation', cartCreate?: { __typename?: 'CartCreatePayload', cart?: (
      { __typename?: 'Cart' }
      & { ' $fragmentRefs'?: { 'CartFragment': CartFragment } }
    ) | null } | null };

export type EditCartItemsMutationVariables = Exact<{
  cartId: Scalars['ID']['input'];
  lines: Array<CartLineUpdateInput> | CartLineUpdateInput;
}>;


export type EditCartItemsMutation = { __typename?: 'Mutation', cartLinesUpdate?: { __typename?: 'CartLinesUpdatePayload', cart?: (
      { __typename?: 'Cart' }
      & { ' $fragmentRefs'?: { 'CartFragment': CartFragment } }
    ) | null } | null };

export type RemoveFromCartMutationVariables = Exact<{
  cartId: Scalars['ID']['input'];
  lineIds: Array<Scalars['ID']['input']> | Scalars['ID']['input'];
}>;


export type RemoveFromCartMutation = { __typename?: 'Mutation', cartLinesRemove?: { __typename?: 'CartLinesRemovePayload', cart?: (
      { __typename?: 'Cart' }
      & { ' $fragmentRefs'?: { 'CartFragment': CartFragment } }
    ) | null } | null };

export type GetCartQueryVariables = Exact<{
  cartId: Scalars['ID']['input'];
  country?: InputMaybe<CountryCode>;
}>;


export type GetCartQuery = { __typename?: 'QueryRoot', cart?: (
    { __typename?: 'Cart' }
    & { ' $fragmentRefs'?: { 'CartFragment': CartFragment } }
  ) | null };

export type CollectionFragment = { __typename?: 'Collection', handle: string, title: string, description: string, updatedAt: any, seo: (
    { __typename?: 'SEO' }
    & { ' $fragmentRefs'?: { 'SeoFragment': SeoFragment } }
  ) } & { ' $fragmentName'?: 'CollectionFragment' };

export type GetCollectionsQueryVariables = Exact<{ [key: string]: never; }>;


export type GetCollectionsQuery = { __typename?: 'QueryRoot', collections: { __typename?: 'CollectionConnection', edges: Array<{ __typename?: 'CollectionEdge', node: { __typename?: 'Collection', title: string, handle: string, description: string, image?: { __typename?: 'Image', url: any, altText?: string | null } | null } }> } };

export type GetMenuQueryVariables = Exact<{
  handle: Scalars['String']['input'];
}>;


export type GetMenuQuery = { __typename?: 'QueryRoot', menu?: { __typename?: 'Menu', title: string, items: Array<{ __typename?: 'MenuItem', title: string, url?: any | null, items: Array<{ __typename?: 'MenuItem', title: string, url?: any | null, items: Array<{ __typename?: 'MenuItem', title: string, url?: any | null }> }> }> } | null };

export type PageFragment = { __typename?: 'Page', id: string, title: string, handle: string, body: any, bodySummary: string, createdAt: any, updatedAt: any, seo?: (
    { __typename?: 'SEO' }
    & { ' $fragmentRefs'?: { 'SeoFragment': SeoFragment } }
  ) | null } & { ' $fragmentName'?: 'PageFragment' };

export type GetPageQueryVariables = Exact<{
  handle: Scalars['String']['input'];
}>;


export type GetPageQuery = { __typename?: 'QueryRoot', pageByHandle?: (
    { __typename?: 'Page' }
    & { ' $fragmentRefs'?: { 'PageFragment': PageFragment } }
  ) | null };

export type GetPagesQueryVariables = Exact<{ [key: string]: never; }>;


export type GetPagesQuery = { __typename?: 'QueryRoot', pages: { __typename?: 'PageConnection', edges: Array<{ __typename?: 'PageEdge', node: (
        { __typename?: 'Page' }
        & { ' $fragmentRefs'?: { 'PageFragment': PageFragment } }
      ) }> } };